remove concludeGameIfOver, move Move types to movepiece.ts
