+++ src/client/scripts/esm/chess/util/moveutil.js
@@ -11,7 +11,7 @@ import coordutil from './coordutil.js';
  * Type Definitions 
  * @typedef {import('../logic/gamefile.js').gamefile} gamefile
  * @typedef {import('../logic/boardchanges.js').Change} Change
- * @typedef {import('../../game/chess/movesequence.js').Move} Move
+ * @typedef {import('../logic/movepiece.js').Move} Move
 */
